Make Target conform ModelConvertible
