+++ Sources/TuistCore/Models/Target.swift
@@ -2,7 +2,7 @@ import Basic
 import Foundation
 import TuistSupport
 
-public class Target: Equatable, Hashable {
+public struct Target: Equatable, Hashable {
     public typealias SourceFile = (path: AbsolutePath, compilerFlags: String?)
     public typealias SourceFileGlob = (glob: String, excluding: String?, compilerFlags: String?)
 
@@ -22,7 +22,7 @@ public class Target: Equatable, Hashable {
 
     // An info.plist file is needed for (dynamic) frameworks, applications and executables
     // however is not needed for other products such as static libraries.
-    public var infoPlist: InfoPlist?
+    public private(set) var infoPlist: InfoPlist?
     public let entitlements: AbsolutePath?
     public let settings: Settings?
     public let dependencies: [Dependency]
@@ -170,6 +170,14 @@ public class Target: Equatable, Hashable {
         hasher.combine(productName)
         hasher.combine(entitlements)
         hasher.combine(environment)
+    }
+
+    /// Returns a copy of the InfoPlist with the given InfoPlist
+    /// - Parameter infoPlist: InfoPlist to be set to the copy.
+    public func with(infoPlist: InfoPlist) -> Target {
+        var target = self
+        target.infoPlist = infoPlist
+        return target
     }
 }
