+++ Sources/TuistLoader/Loaders/GeneratorModelLoader.swift
@@ -190,9 +190,7 @@ extension TuistCore.Project {
 
         let settings = try manifest.settings.map { try TuistCore.Settings(manifest: $0, generatorPaths: generatorPaths) }
         let targets = try manifest.targets.map {
-            try TuistCore.Target.from(manifest: $0,
-                                      path: path,
-                                      generatorPaths: generatorPaths)
+            try TuistCore.Target(manifest: $0, generatorPaths: generatorPaths)
         }
 
         let schemes = try manifest.schemes.map { try TuistCore.Scheme(manifest: $0, generatorPaths: generatorPaths) }
@@ -239,69 +237,5 @@ extension TuistCore.Project {
                 packages: packages,
                 schemes: schemes,
                 additionalFiles: additionalFiles)
-    }
-}
-
-extension TuistCore.Target {
-    // swiftlint:disable:next function_body_length
-    static func from(manifest: ProjectDescription.Target,
-                     path: AbsolutePath,
-                     generatorPaths: GeneratorPaths) throws -> TuistCore.Target {
-        let name = manifest.name
-        let platform = try TuistCore.Platform(manifest: manifest.platform, generatorPaths: generatorPaths)
-        let product = try TuistCore.Product(manifest: manifest.product, generatorPaths: generatorPaths)
-
-        let bundleId = manifest.bundleId
-        let productName = manifest.productName
-        let deploymentTarget = try manifest.deploymentTarget.map { try TuistCore.DeploymentTarget(manifest: $0, generatorPaths: generatorPaths) }
-
-        let dependencies = try manifest.dependencies.map { try TuistCore.Dependency(manifest: $0, generatorPaths: generatorPaths) }
-
-        let infoPlist = try TuistCore.InfoPlist(manifest: manifest.infoPlist, generatorPaths: generatorPaths)
-        let entitlements = try manifest.entitlements.map { try generatorPaths.resolve(path: $0) }
-
-        let settings = try manifest.settings.map { try TuistCore.Settings(manifest: $0, generatorPaths: generatorPaths) }
-        let sources = try TuistCore.Target.sources(projectPath: path, sources: manifest.sources?.globs.map {
-            let glob = try generatorPaths.resolve(path: $0.glob).pathString
-            let excluding = try $0.excluding.map { try generatorPaths.resolve(path: $0).pathString }
-            return (glob: glob, excluding: excluding, compilerFlags: $0.compilerFlags)
-        } ?? [])
-
-        let resourceFilter = { (path: AbsolutePath) -> Bool in
-            TuistCore.Target.isResource(path: path)
-        }
-        let resources = try (manifest.resources ?? []).flatMap {
-            try TuistCore.FileElement.from(manifest: $0,
-                                           path: path,
-                                           generatorPaths: generatorPaths,
-                                           includeFiles: resourceFilter)
-        }
-
-        let headers = try manifest.headers.map { try TuistCore.Headers(manifest: $0, generatorPaths: generatorPaths) }
-
-        let coreDataModels = try manifest.coreDataModels.map {
-            try TuistCore.CoreDataModel(manifest: $0, generatorPaths: generatorPaths)
-        }
-
-        let actions = try manifest.actions.map { try TuistCore.TargetAction(manifest: $0, generatorPaths: generatorPaths) }
-        let environment = manifest.environment
-
-        return TuistCore.Target(name: name,
-                                platform: platform,
-                                product: product,
-                                productName: productName,
-                                bundleId: bundleId,
-                                deploymentTarget: deploymentTarget,
-                                infoPlist: infoPlist,
-                                entitlements: entitlements,
-                                settings: settings,
-                                sources: sources,
-                                resources: resources,
-                                headers: headers,
-                                coreDataModels: coreDataModels,
-                                actions: actions,
-                                environment: environment,
-                                filesGroup: .group(name: "Project"),
-                                dependencies: dependencies)
     }
 }
